% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_gwas_aux.R
\name{asrgen.args}
\alias{asrgen.args}
\title{\pkg{ASRgenomics} function arguments for \link{pre.gwas}}
\usage{
asrgen.args()
}
\description{
\pkg{ASRgenomics} function arguments for \link{pre.gwas}
}
\keyword{internal}
