% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_generic.R
\name{audit.generic}
\alias{audit.generic}
\title{Checks/audits datasets and matrices prior to GBLUP/GWAS}
\usage{
audit.generic(
  pheno.data = NULL,
  indiv = NULL,
  resp = NULL,
  Kinv = NULL,
  geno.data = NULL,
  map.data = NULL,
  Q = NULL,
  caller = c("gwas", "gblup"),
  message = TRUE
)
}
\description{
See \link[ASRgblup]{audit.gblup} and \link[ASRgwas]{audit.gwas}
}
\keyword{internal}
