% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_elimination.R
\name{backward.elimination}
\alias{backward.elimination}
\title{Implementation of GWAS model selection by backward elimination}
\usage{
backward.elimination(
  mod = NULL,
  data = NULL,
  try.variables = NULL,
  pvalue.thr = 0.1,
  maxiter.update = 3,
  workspace = "1Gb",
  message = TRUE
)
}
\arguments{
\item{mod}{A pre-fit \pkg{asreml} model object. Must follow the conditions of the function
(default = \code{NULL}).}

\item{data}{The data used in the process of fitting the \code{asreml} model with additional
variables to be tested as fixed effects (default = \code{NULL}).}

\item{try.variables}{A character vector with the name of variables from \code{data} to be
tested in the backward selection procedure (default = \code{NULL}).
#' @param pvalue.thr A numeric value with \eqn{p-value} threshold to identify significant markers
for the elimination process. Markers will be eliminated until all have a \eqn{p-value} smaller
than the value provided (default = \code{0.1})}

\item{maxiter.update}{A numeric value indicating the number of iterations to be
used in \link[asreml]{update.asreml} (default = \code{3}).}

\item{workspace}{Specifies the workspace to be used by the \code{asreml} function
(default = \code{"1Gb"}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A list containing:

\itemize{
  \item \code{mod}: The updated base model fitted with the retained variables.
  \item \code{kept.variables}: A vector with the names of variables with \eqn{p}-value < \code{pvalue.thr}.
  \item \code{aov}: The ANOVA table of the updated model with the kept variables.
}
}
\description{
Performs backward elimination over a set of provided variables based on their
\eqn{p-value} when added to a reference model.
}
\details{
The elimination technique starts by adding all \code{try.variables} to the base model
as fixed effects and calculating their \eqn{p-value}. The variable with the highest \eqn{p-value}
is eliminated and the model is re-fitted. The process is repeated until all retained variables
have a \eqn{p-value} larger than the provided \code{pvalue.thr}.

If \code{vm()} was used to fit the base model, the relationship matrix used \strong{must} be
available in \code{.GlobalEnv} under the same name as used in \code{asreml()}.
}
\examples{
\dontrun{
model <- asreml::asreml(yield ~ Variety*Nitrogen, random = ~ Blocks/Wplots,
  na.action = list(y='omit', x='omit'), data = oats)

bwd.sel <- backward.elimination(
  mod = model, data = oats, try.variables = c("Row", "Column"),
  pvalue.thr = 0.1, maxiter.update = 10)

bwd.sel <- backward.elimination(
  mod = model, data = oats,  try.variables = c("Row", "Column"),
  pvalue.thr = 0.1, maxiter.update = 10)

bwd.sel$aov
bwd.sel$kept.variables
}
}
\keyword{internal}
