% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_generator.R
\name{call.generator}
\alias{call.generator}
\title{Simple argument-oriented function to build \pkg{asreml} calls}
\usage{
call.generator(
  resp = NULL,
  fixedf = NULL,
  group.cols = NULL,
  cov = NULL,
  randomf = NULL,
  randomf.vm = NULL,
  vm.matrix.name = NULL,
  data.name = NULL,
  residual = NULL,
  weights = NULL,
  family = NULL,
  dispersion = 1,
  total = NULL,
  workspace = NULL
)
}
\arguments{
\item{resp}{A character with name to be used as response variable in \pkg{asreml} call (default = \code{NULL}).}

\item{fixedf}{A character vector with variable names to be used as fixed effects in \pkg{asreml} call.
Interactions/nesting are allowed (default = \code{NULL}).}

\item{group.cols}{A numeric vector with the number of the column(s) in data to be grouped with the
\code{grp()} function (default = \code{NULL}).}

\item{cov}{A character vector with variable names to be used as covariates in \pkg{asreml} call
(default = \code{NULL}).}

\item{randomf}{A character vector with variable names to be used as random effects in \pkg{asreml} call.
Interactions/nesting are allowed (including with randomf.vm) (default = \code{NULL}).}

\item{randomf.vm}{A character vector with variable names to be used as random effects with
user defined relationship matrix in \pkg{asreml} call.
Interactions/nesting NOT allowed. Specify all interactions in \code{randomf} (default = \code{NULL}).}

\item{vm.matrix.name}{The name of the object containing the relationship matrix to be passed to
\code{vm()} (default = \code{NULL}).}

\item{data.name}{A character to be used as name of the data frame in the \pkg{asreml} call (default = \code{NULL}).}

\item{residual}{A character with the name of a variable in \code{pheno.data} to be used as
conditional factor for specifying heterogeneous residuals by level of this factor.}

\item{weights}{A character with name of the numeric variable in \code{pheno.data} with the
weights of each prediction (\emph{i.e.}, mean estimate) (default = \code{NULL}).
Note that \code{pheno.data} will be internally ordered by this factor (default = \code{NULL}).}

\item{family}{A character indicating the family of distribution for \code{resp}.
Options are: \code{"gaussian"} and \code{"binomial"} (default = \code{"gaussian"}).}
}
\value{
A character string with the \pkg{asreml} call.
}
\description{
Simple argument-oriented function to build \pkg{asreml} calls
}
\details{
Currently available models in this function can be built with:
any combination of fixed and random effects;
covariates; groups (\code{grp()}); user defined structures; \code{vm():ind()} interactions/nesting terms;
heterogeneous residuals with \code{dsum(|)}; addition of weights; Gaussian and binomial families.
}
\examples{
\dontrun{
# A dummy example.
asreml.call <-
  ASRgwas:::call.generator(asreml.obj.name = "mod",
    resp = "resp", fixedf = c("year", "block"), group.cols = c(10,12),
    cov = "stand", randomf = c("gen:year"), randomf.vm = "gen",
    vm.matrix.name = "Kinverse", data.name = "phenotypic.data",
    residual = "year", weights = NULL, family = "gaussian", workspace = "4Gb")
str2lang(asreml.call)

# An (adapted) example from asreml-r manual.
asreml.call <-
  ASRgwas:::call.generator(
    asreml.obj.name = "mod",
    resp = "yield", fixedf = c("Blocks", "Nitrogen"),
    cov = "Column", randomf = c("Variety", "Subplots"),
    data.name = "oats", family = "gaussian", workspace = "1Gb")
eval(parse(text = asreml.call))
}
}
\keyword{internal}
