% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_aux.R
\name{check.special}
\alias{check.special}
\title{Identifies the structure (correlation type) and heterogeneity/homogeneity of variance [adapted from asremlPlus]}
\usage{
check.special(var, term, G.param, specials, residual = FALSE)
}
\arguments{
\item{var}{A character with the name of the factor inside the current term (\emph{e.g.}, \code{family}) (no default).}

\item{term}{A character with the name of the term (single factor or combination of factors) in
\code{random=~} (\emph{e.g.}, \code{family:vm(fishid, Kinv.tmp$fishid)}) (no default).}

\item{G.param}{The full object named \code{G.param} from the object of class \code{asreml} (no default).}

\item{specials}{A character vector with the structures allowed (\emph{e.g.}, \code{c("id", "diag")}) (no default).}

\item{residual}{If \code{TRUE} it means we are working with \eqn{\strong{R}} structures (default = \code{FALSE}).}
}
\value{
A list with the structure name and the final (\code{"v"} or \code{"h"}).
The output is required by \link{g.asreml} for the reproduction of the structures in matrix form.
}
\description{
Identifies the structure (correlation type) and heterogeneity/homogeneity of variance [adapted from asremlPlus]
}
\details{
It will identify the correlation structure (\emph{e.g.}, \code{"id"}) of the current var within term.
Also, will identify if the structure has homogeneous or heterogeneous variance components based
on the end of the structure, which can be \code{"v"} or \code{"h"}.
}
\author{
Chris Brien; VSNi
}
\keyword{internal}
