% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{correlation.report_}
\alias{correlation.report_}
\title{Identifies high correlations in matrix of data}
\usage{
correlation.report_(matrix_ = NULL, cor.thr_ = 0.95)
}
\arguments{
\item{matrix_}{An object of class matrix to be checked for high correlations (default = \code{NULL}).}

\item{cor.thr_}{A numeric value with the threshold value to generate warnings (default = 0.95).}
}
\value{
Returns a \code{data.frame} with highly correlated combinations.
}
\description{
Identifies high correlations in matrix of data
}
\keyword{internal}
