% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{ellipsis.subset}
\alias{ellipsis.subset}
\title{Subset ellipsis based on a functions formals.}
\usage{
ellipsis.subset(ellipsis = NULL, fun = NULL)
}
\arguments{
\item{ellipsis}{The ellipsis (\emph{i.e.}, ...) (default = \code{NULL}).}

\item{fun}{The function to check formals and collect respective input from ... (default = \code{NULL}).}
}
\value{
A list with the ... corresponding to the function.
}
\description{
Sometimes we want the ellipsis to bear the input for more than one internal function.
Therefore we can split it according to the formals of each function.
}
\keyword{internal}
