% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{fix.call.data_}
\alias{fix.call.data_}
\title{Changes data argument in \pkg{asreml} call}
\usage{
fix.call.data_(mod_ = NULL, replacement_ = NULL)
}
\arguments{
\item{mod_}{An \pkg{asreml} object (default = \code{NULL}).}
}
\value{
The call with the adjusted values in the \code{data} argument.
}
\description{
This function will change the call by adjusting the name to be used in
the \code{data} argument.
}
\keyword{internal}
