% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{fix.call.grp_}
\alias{fix.call.grp_}
\title{Change group argument \pkg{asreml} call to match \code{.$mf}}
\usage{
fix.call.grp_(mod_ = NULL)
}
\arguments{
\item{mod_}{An \pkg{asreml} object that used the \code{group} argument (default = \code{NULL}).}
}
\value{
The call with the adjusted values in the \code{group} argument.
}
\description{
This function will correct the call by adjusting the variables to be used in
the \code{group} argument. The positions might change from the used data and the data
reported in \code{.$mf}.
}
\keyword{internal}
