% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{fix.call.param_}
\alias{fix.call.param_}
\title{Removes G.param and R.param arguments from \pkg{asreml} object call}
\usage{
fix.call.param_(mod_ = NULL)
}
\arguments{
\item{mod_}{An \strong{updated} (with \link[asreml]{update.asreml}) \pkg{asreml} object (default = \code{NULL}).}
}
\value{
The call without the \code{G.param} and \code{R.param} arguments.
}
\description{
Removes G.param and R.param arguments from \pkg{asreml} object call
}
\keyword{internal}
