% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_aux.R
\name{g.asreml}
\alias{g.asreml}
\title{Gets incidence matrices to estimate \eqn{\strong{G}} [adapted from asremlPlus]}
\usage{
g.asreml(var, term, G.param, kspecial, cond.fac = "")
}
\arguments{
\item{var}{A character with the name of the factor inside the current term (\emph{e.g.}, \code{family}) (no default).}

\item{term}{A character with the name of the term (single factor or combination of factors) in
\code{random=~} (\emph{e.g.}, \code{family:vm(fishid, Kinv.tmp$fishid)}) (no default).}

\item{G.param}{The full object named \code{G.param} from the object of class \code{asreml} (no default).}

\item{kspecial}{The output from the \code{check.special} function (no default).}

\item{cond.fac}{NOT REQUIRED FOR THE CURRENT ACCEPTED CORRELATION STRUCTURES.}
}
\value{
A "correlation matrix" or \eqn{\strong{G}} (if \code{$model} is \code{idv}) associated to a given var (factor).
}
\description{
Collects information from \code{G.param} and produce the correlation
matrix used to get \eqn{\strong{G}}.
}
\details{
If \code{$variance$model == "id"} it means that the variance component is in the
\code{$-name of var-$initial} object. In this case the correlation matrix is multiplied by
the variance to get \eqn{\strong{G}} already (otherwise it is done outside the function). This
happens, for example, in interactions.

For now, only \code{id} correlation structures is accepted, others can be added by un-commenting
the commented lines (more has to be changed outside).
}
\author{
Chris Brien; VSNi
}
\keyword{internal}
