% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_aux.R
\name{G.id}
\alias{G.id}
\title{Gets a diagonal matrix based on the number of levels of var within term [adapted from asremlPlus]}
\usage{
G.id(var, term, G.param, cond.fac = "")
}
\arguments{
\item{var}{A character with the name of the factor inside the current term (\emph{e.g.}, \code{family}) (no default).}

\item{term}{A character with the name of the term (single factor or combination of factors) in
\code{random=~} (\emph{e.g.}, \code{family:vm(fishid, Kinv.tmp$fishid)}) (no default).}

\item{G.param}{The full object named \code{G.param} from the object of class \code{asreml} (no default).}

\item{cond.fac}{Placeholder. Switched off in this function (no default).}
}
\value{
A diagonal matrix.
}
\description{
Gets a diagonal matrix based on the number of levels of var within term [adapted from asremlPlus]
}
\author{
Chris Brien; VSNi
}
\keyword{internal}
