% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{get.vm.termvar}
\alias{get.vm.termvar}
\title{Gets the term name for vms (with relationship matrix)}
\usage{
get.vm.termvar(factor = NULL, source = "Kinv")
}
\arguments{
\item{factor}{A character vector to be placed in vm(*here*, source$*here*) (default = \code{NULL}).}

\item{source}{A character vector to be placed in vm(term, *here*$term) (default = \code{"Kinv"}).}
}
\value{
A character vector, \emph{e.g.}, \code{"vm(genA, Kinv$genA)" "vm(genD, Kinv$genD)"}
}
\description{
Gets the term name for vms (with relationship matrix)
}
\examples{
ASRgwas:::get.vm.termvar(c("genA", "genD"), source = "gwas.data$Kinv")
}
\keyword{internal}
