% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_asreml_core.R
\name{gwas.asreml.core}
\alias{gwas.asreml.core}
\title{Directs data frames and matrices to the respective GWAS method}
\usage{
gwas.asreml.core(
  mod = NULL,
  weights = NULL,
  map.data = NULL,
  geno.data = NULL,
  Kinv = NULL,
  pvalue.thr = 5e-06,
  P3D = TRUE,
  user.mod = NULL,
  maxiter.update = 1,
  cpp = NULL,
  inverse.update = NULL,
  threads = 1,
  obj.parallel = NULL,
  message = TRUE
)
}
\arguments{
\item{mod}{A pre-fit \pkg{asreml} model object. Must follow the conditions of the function
(default = \code{NULL}).}

\item{weights}{A character with name of the numeric variable in \code{pheno.data} with the
weights of each prediction (\emph{i.e.}, mean estimate) (default = \code{NULL}).}

\item{map.data}{A data frame with each marker's name, chromosome, and position.
Variable names \strong{must} be "marker", "chrom", and "pos" (default = \code{NULL}).}

\item{geno.data}{A matrix with SNP data of form \eqn{n \times p},
with \eqn{n} individuals and \eqn{p} markers. Individual and marker names are
assigned to \code{rownames} and \code{colnames}, respectively.
SNP data is coded as: 0, 1, 2 (default = \code{NULL}).}

\item{Kinv}{A list of matrices representing the \strong{inverse} of the genomic relationship
matrix \eqn{\boldsymbol{Kinv}}. The matrices must be in sparse form (default = \code{NULL}).}

\item{pvalue.thr}{A numeric value with \eqn{p-value} threshold to identify significant markers (default = \code{5e-6})}

\item{P3D}{If \code{TRUE} the "Population Parameters Previously Determined" algorithm will be used.
If \code{FALSE} the variance components are allowed to vary while fitting each marker individually.
In setting \code{P3D = TRUE} there is a considerable speed gain for \code{family = "gaussian"}, but not for
\code{family = "binomial"} (default = \code{TRUE}).}

\item{user.mod}{Set to \code{TRUE} if the base model has been calculated outside \link{gwas.asreml}
(default = \code{NULL}).}

\item{maxiter.update}{A numeric value indicating the number of iterations to be
used in \link[asreml]{update.asreml} under the non-P3D method (default = \code{1}).}

\item{threads}{An integer with the number of threads to be used in parallel processing
(default = \code{1}).}

\item{obj.parallel}{A character vector indicating which objects passed to the function that must
be exported to clusters (\emph{e.g.}, objects passed to \code{Kinv} argument). If nothing is passed, the
algorithm will try to identify the required objects (default = \code{NULL}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A list with several outputs from the GWAS model fit that will be passed to \link{gwas.asreml}:

\itemize{
  \item \code{call}: The \pkg{asreml} model call.
  \item \code{gwas.all}: A data frame with map and marker statistics of all tested markers.
  \item \code{gwas.sel}: A data frame with map and marker statistics of significant markers.
  \item \code{mod}: The \pkg{asreml} object GWAS model fitted with the provided data.
}
}
\description{
Passes the data frames and matrices to the available GWAS model fit functions.
Currently methods: \code{P3D} and \code{marker.update} are available.
}
\keyword{internal}
