% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_asreml_main.R
\name{gwas.asreml}
\alias{gwas.asreml}
\title{Fits Genome-wide Association Studies (GWAS) models using asreml}
\usage{
gwas.asreml(
  pheno.data = NULL,
  resp = NULL,
  gen = NULL,
  Kinv = NULL,
  Q = NULL,
  npc = 10L,
  cov = NULL,
  fixedf = NULL,
  randomf = NULL,
  residual = NULL,
  weights = NULL,
  family = c("gaussian", "binomial"),
  dispersion = 1,
  total = NULL,
  workspace = "1Gb",
  mod = NULL,
  geno.data = NULL,
  map.data = NULL,
  pvalue.thr = 0.05,
  bonferroni = TRUE,
  P3D = TRUE,
  maxiter.update = 2L,
  inverse.update = -1L,
  threads = detectCores() - 1,
  obj.parallel = NULL,
  message = TRUE
)
}
\arguments{
\item{pheno.data}{A data frame with all relevant columns (fphenotypic response,
factors and covariates) to be used (default = \code{NULL}).}

\item{resp}{A character string of the name of the numeric variable in \code{pheno.data}
that identifies the response variable (default = \code{NULL}).}

\item{gen}{A character (vector) with the names of the columns in \code{pheno.data}
that identifies the genotypes for the "genetic" part of the model (default = \code{NULL}).}

\item{Kinv}{A list of matrices representing the \strong{inverse} of the genomic relationship
matrix \eqn{\boldsymbol{Kinv}}. The matrices must be in sparse form (default = \code{NULL}).}

\item{Q}{A matrix of size \eqn{n} x \eqn{nv}, with \eqn{nv} population structure-related covariates (\emph{e.g.},
principal component scores). The number of components to be included in the fitted model
is determined by \code{npc} (default = \code{NULL}).}

\item{npc}{The number of components (columns) to be used from the \eqn{\boldsymbol{Q}} matrix
(default = \code{10L}).}

\item{cov}{A character vector with the names of the numeric variables in \code{pheno.data}
to be considered as covariates in the GWAS model (default = \code{NULL}).}

\item{fixedf}{A character vector with the names of the factors in \code{pheno.data}
to be considered as fixed effects in the GWAS model. Interactions can be added as \code{"factor1:factor2"}
(default = \code{NULL}).}

\item{randomf}{A character vector with the names of the factors in \code{pheno.data}
to be considered as random effects in the GWAS model. Interactions can be added as
\code{"factor1:factor2"}. Interactions with \code{gen} factor can also
be added here (\emph{e.g.}, \code{"gen:factor1"}) (random by interaction with \code{gen}).
Genotypes should be added to the argument \code{gen}, not the \code{randomf} (default = \code{NULL}).}

\item{residual}{A character string of the name of the variable in \code{pheno.data} to be used as a
conditional factor for specifying a heterogeneous residual structure according to the levels of this factor.
The \code{pheno.data} will be internally ordered by this factor (default = \code{NULL}).}

\item{weights}{A character stirng of the name of the numeric variable in \code{pheno.data} that
specifics the weights of each prediction (\emph{i.e.}, mean estimate) (default = \code{NULL}).}

\item{family}{A character indicating the family of the distribution for \code{resp}.
Options are: \code{"gaussian"} and \code{"binomial"} (default = \code{"gaussian"}).}

\item{dispersion}{A numeric value of the dispersion parameter
(used if \code{family = "binomial"}) (default = \code{1}).}

\item{total}{A character string of the name of the numeric variable in \code{pheno.data} that
indicates the total counts (if \code{family = "binomial"}). If \code{NULL}, count is taken as 1
(default = \code{NULL}).}

\item{workspace}{Specifies the workspace to be used by the \code{asreml} function.
Note that large workspaces will result in much longer processing times if \code{P3D} is false
(default = \code{"1Gb"}).}

\item{mod}{A pre-fitted \pkg{asreml} model object. Providing a
pre-fitted model works for Gaussian
data if \code{P3D = FALSE} and binomial data (default = \code{NULL}).}

\item{geno.data}{A matrix with SNP data of form \eqn{n \times p},
with \eqn{n} individuals and \eqn{p} markers. Individual and marker names are
assigned to \code{rownames} and \code{colnames}, respectively.
SNP data is coded as: 0, 1, 2 (default = \code{NULL}).}

\item{map.data}{A data frame with the marker names, chromosomes, and positions.
Variable names \strong{must} be: "marker", "chrom", and "pos" (default = \code{NULL}).}

\item{pvalue.thr}{A numeric value with the \eqn{p-value} threshold to identify significant markers (default = \code{0.05})}

\item{bonferroni}{If \code{TRUE} the Bonferroni correction is used with expression \eqn{1 - (1 - pvalue.thr)^{1/p} \sim pvalue.thr/p},
where \eqn{pvalue.thr} is based on the argument (\code{pvalue.thr}) and \eqn{p} is the number of markers (default = \code{TRUE}).}

\item{P3D}{If \code{TRUE} the "Population Parameters Previously Determined" algorithm will be used.
If \code{FALSE} the variance components are allowed to vary while fitting each marker individually.
Setting \code{P3D = TRUE} results in considerable speed gain for \code{family = "gaussian"}, but not for
\code{family = "binomial"} (default = \code{TRUE}).}

\item{maxiter.update}{A numeric value indicating the number of iterations to be
used in \code{update.asreml} method if requested (default = \code{2}).}

\item{inverse.update}{An integer indicating the number of Schulz-type iterative updates to be applied to
the inverse of \eqn{\boldsymbol{V}} (phenotypic variance matrix) if \code{geno.data}
has missing values when running a Gaussian model under P3D.
The larger the value the better the approximation (under certain conditions).
If \code{inverse.update = -1} then the Woodbury's method is used instead.
If \code{inverse.update = 0} is used, then Schulz's procedure is performed with no iterations (no update),
which is less precise but faster to run. See details for more information (default = \code{-1}).}

\item{threads}{An integer with the number of thread to be used in parallel processing
(default = \code{parallel::detectCores() - 1}; \emph{i.e.}, total number of threads available minus one).}

\item{obj.parallel}{A character vector indicating which objects to be passed to the function that must
be exported to clusters (\emph{e.g.} objects passed to \code{Kinv} argument). If nothing is passed, the
algorithm will try to identify the required objects (default = \code{NULL}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A list of class \code{gwas.asreml} with several outputs from a GWAS model fit:

\itemize{
  \item \code{call}: An object of class \code{call} with the code used to fit \code{mod} in \pkg{asreml}.
  \item \code{gwas.all}: A data frame with map (\code{marker}, \code{chrom}, \code{pos}),
    minor allele frequency (\code{maf}), \code{effect}, \code{z.ratio}, \code{p.value},
    and explained variance (\code{expl.var}; see details) of all tested markers.
  \item \code{gwas.sel}: A subset of \code{gwas.all} with significant markers based on \code{pvalue.thr}
    and/or Bonferroni.
  \item \code{mod}: The \pkg{asreml} object base model fitted with the provided data.
  \item \code{aov}: The ANOVA table of the fitted base model.
  \item \code{heritability}: The estimated heritability value for \code{h2.vc} and
    \code{h2.pev} definitions.
}
}
\description{
Main function of \pkg{ASRgwas} that extends
the capabilities of \pkg{asreml} to perform Genome-wide Association Studies.
}
\details{
Some relevant considerations are presented below.

Marker only models (no \eqn{\boldsymbol{Q}} or \eqn{\boldsymbol{K}}) can be fitted by specifying
\code{gen}, \code{Q}, and \code{Kinv} as \code{NULL}.

The factors passed to \code{gen} are considered random, as are the ones passed via \code{randomf}.
The arguments are separated so it is clear which ones should be on the numerator (\code{gen})
and denominator (\code{gen + randomf}) of the heritability (or repeatability) formula.
Interactions with \code{gen} factors should be included in \code{randomf} and will be
always added to the denominator. No interactions between \code{gen} factors are allowed.

If there is more than one type of genetic effect to be modeled (but associated
with different genomic matrices),
the column with genotype identifications in \code{pheno.data} should be copied with a
different name so there is one for each genetic effect considered.

If relationship matrices are associated to \code{gen} factors,
they should be passed using \code{Kinv}. The name of objects in \code{Kinv} \strong{must}
match the values passed to \code{gen}. For example, to model additive and dominance effects,
the following code can be used: \code{gen = c("genA", "genD")}, where \code{"genA"} and
\code{"genD"} are columns in \code{pheno.data} with genotype identifiers;
and \code{Kinv = list("genA" = KinvA, "genD" = KinvD)}, where \code{KinvA} and
\code{KinvD} are inverse matrices of additive and dominance relationships in sparse form.
If a factor is passed to \code{gen} but not to \code{Kinv},
it is regarded as having independent levels (\emph{i.e.}, \code{id/idv} variance structure).
Also, all factors names passed to the command \code{Kinv} \strong{must} be present in
the command \code{gen}.

Row and column names are checked for matching between and within lists, and between datasets.
In the case of the sparse form \code{Kinv}, attribute names \code{rowNames} and \code{colNames}
are used to perform this verification.

Any number of fixed (factors and covariates) and random effects can be passed,
but only single trait analysis has been implemented so far.

Missing values in \code{resp} and \code{weights} are allowed. They are dealt with by
eliminating the observations (rows) prior to the function call.

Missing values on the marker are allowed in both \code{P3D = TRUE} and \code{P3D = FALSE}.
If \code{P3D = FALSE}, \pkg{asreml} deals with the missing data.
Two algorithms (Schulz and Woodbury) have been implemented to deal
with missing data by adjusting the inverse of the \eqn{\boldsymbol{V}}
(phenotypic variance) matrix if \code{P3D = TRUE}.
It is likely that Woodbury's method will be faster and more precise than
Schulz's method with one or more iterations. In addition, Schulz's method
is highly dependent on matrix stability, while Woodbury's method is more robust.
Note that the default is Woodbury's method.

The variance explained by each marker present in \code{gwas.all/gwas.sel}
is calculated via \eqn{expl.var = 100 \times [2 \times maf \times (1-maf) \times effect^2] / var(resp)}.
The MAF (\code{maf}; and consequently the variance explained \code{expl.var}) are only correct
for \code{geno.data} with additive genotype coding.

The calculation of the \eqn{\boldsymbol{V}} (phenotypic variance) matrix required by Gaussian P3D
is based on the code previously implemented for the library \pkg{asremlPlus} (Brien, 2021).
}
\examples{
\dontrun{
# Prepare Apricot dataset.
gwas.data <- pre.gwas(
  pheno.data = pheno.apricot, indiv = "Ind", resp = "Sucrose",
  geno.data = geno.apricot, map.data = map.apricot,
  maf = 0.05, heterozygosity = 0.9, Fis = 0.5,
  marker.callrate = 0.40, impute = TRUE)

# Fit model to see how good it is.
gwasA <- gwas.asreml(
  pheno.data = gwas.data$pheno.data,
  resp = "Sucrose",
  gen = "Ind",
  fixedf = "Lots",
  residual = "Lots",
  Kinv = gwas.data$Kinv,
  Q = gwas.data$Q,
  npc = 3)

# Check the call.
gwasA$call

# Check heritability/repeatability.
gwasA$heritability

# Check the ANOVA table.
gwasA$aov

# Fit the GWAS for that model.
gwasA <- gwas.asreml(
  pheno.data = gwas.data$pheno.data,
  resp = "Sucrose",
  gen = "Ind",
  fixedf = "Lots",
  residual = "Lots",
  Kinv = gwas.data$Kinv,
  Q = gwas.data$Q,
  npc = 3,
  geno.data = gwas.data$geno.data,
  map.data = gwas.data$map.data,
  pvalue.thr = 0.0005,
  bonferroni = FALSE,
  workspace = "2Gb")

# The quantile-quantile plot.
qq.plot(gwas.table = gwasA$gwas.all)

# The Manhattan plot.
manhattan.plot(gwas.table = gwasA$gwas.all, padding = 2e6)

# Statistics for all markers.
head(gwasA$gwas.all)

# Statistics for only significant markers.
head(gwasA$gwas.sel)
}
}
\references{
Brien C. 2021.
asremlPlus: Augments 'ASReml-R' in fitting mixed models and packages generally in exploring prediction differences.
R package version 4.3-31, <https://CRAN.R-project.org/package=asremlPlus>.

Hager, W.W. 1989. Updating the inverse of a matrix.
\emph{SIAM Review} 31(2):221–239.

Petković M.D. 2014.
Generalized Schultz iterative methods for the computation of outer inverses.
\emph{Computers & Mathematics with Applications} 67(10).

Zhang, Z. et al. 2010. Mixed linear model approach adapted for genome-wide association studies.
\emph{Nature Genetics}, 42(4):355–360.
}
