% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_from_asreml.R
\name{input.from.asreml}
\alias{input.from.asreml}
\title{Collects data from an asreml object}
\usage{
input.from.asreml(mod = NULL, parent.frame = 1)
}
\arguments{
\item{mod}{An object of class \code{asreml} (default = \code{NULL}).}

\item{parent.frame}{The environment in which the data will be assigned (default = \code{1} which
means one environment above).}
}
\value{
The \code{pheno.data}, \code{resp}, \code{family}, \code{weight}, \code{residual}
on the selected environment.
}
\description{
Collects data from an asreml object
}
\keyword{internal}
