% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan.plot}
\alias{manhattan.plot}
\title{Draws one or more Manhattan plots for GWAS marker data results}
\usage{
manhattan.plot(
  gwas.table = NULL,
  pvalue.thr = NULL,
  tag.table = NULL,
  tag.repel = TRUE,
  point.colour = "chrom",
  point.size = 1,
  point.alpha = 1,
  collate = TRUE,
  padding = 0,
  gwas.index = NULL,
  facet.main = NULL,
  facet.secondary = NULL,
  facet.dim = NULL,
  scales = "free_y",
  legend.position = "none",
  message = TRUE
)
}
\arguments{
\item{gwas.table}{A data frame of one or more (vertically stacked) GWAS analysis with (as a minimum) the
following variables: \code{"marker"}, \code{"chrom"}, \code{"pos"},
and \code{"p.value"} of markers (default = \code{NULL}).}

\item{pvalue.thr}{A numeric value with the \code{p-value} threshold to identify significant markers (\emph{e.g.}, \code{5e-4}).
The value will be transformed using \code{-log10(pvalue.thr)} for plotting (default = \code{NULL}).}

\item{tag.table}{A data frame of tags (annotations) for scores. The columns \code{"marker"} and
\code{"tag"} are required. The column \code{"tag"} should contain the
text to be annotated in each marker score (\emph{e.g.}, "Candidate region 3"). Columns
\code{"nudge.x"} and \code{"nudge.y"} are optional to inform position nudges
on the text and should be on the base pair number and \code{log(p-value)} scale,
respectively (default = \code{NULL}).}

\item{tag.repel}{If \code{TRUE}, \code{nudge.x} and \code{nudge.y} from \code{tag.table} are ignored
and tags are automatically repelled (default = \code{TRUE}).}

\item{point.colour}{A character with a colour name (\emph{e.g.}, \code{"blue"}) or a column name
from \code{gwas.table} to be used as the index for point colouring (default = \code{"chrom"}).}

\item{point.size}{A numeric value (greater than 0) indicating the point size
(default = \code{1}).}

\item{point.alpha}{A numeric value indicating the points' transparency (default = \code{1}).}

\item{collate}{If \code{TRUE} the distance between the last marker of a chromosome and
the first marker of the next chromosome is set to 1 base pair (default = \code{TRUE}).}

\item{padding}{A numeric value that will be used as the distance between chromosomes (default = \code{0}).}

\item{gwas.index}{A character indicating a column name in \code{gwas.table} identifying each GWAS analysis.
This is only necessary when more than one GWAS analysis is present in \code{gwas.table} (default = \code{NULL}).}

\item{facet.main}{A character indicating a column name in \code{gwas.table} to be used
for the main faceting of panels (if more than one GWAS analysis is present) (default = \code{NULL}).}

\item{facet.secondary}{A character indicating a column name in \code{gwas.table} to be used
as secondary faceting of panels (if more than one GWAS analysis is present) (default = \code{NULL}).}

\item{facet.dim}{A numeric vector indicating the number of rows and columns to be used
in the faceting process, \emph{e.g.}, \code{c(4, 1)}, for 4 rows and 1 column. This argument only works
if \code{facet.main} is provided, and works best if \code{facet.secondary} is \code{NULL}
(default = \code{NULL}).}

\item{scales}{A character indicating if the scales/levels in \eqn{x} and \eqn{y} axis should be the
same values across facets (panels). Options are: \code{"fixed"} \code{"free_x"}, \code{"free_y"},
and \code{"free"}. See \link[ggplot2]{facet_wrap} for more information
(default = \code{"free_y"}).}

\item{legend.position}{A character indicating the position for the legend in the displayed plot.
Options are: \code{"bottom"}, \code{"top"}, \code{"left"}, \code{"right"}, \code{"none"}, or the
corresponding vector of coordinates (such as: \code{c(0.95, 0.05)} for \eqn{x} and \eqn{y}, respectively)
(default = \code{"none"}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A ggplot object containing one or more Manhattan plots.
}
\description{
Generates one or more Manhattan plots for visualization.
}
\examples{
\dontrun{
# Prepare Apricot dataset.
gwas.data <- pre.gwas(
  pheno.data = pheno.apricot, indiv = "Ind", resp = "Sucrose",
  geno.data = geno.apricot, map.data = map.apricot, maf = 0.05,
  marker.callrate = 0.40, impute = TRUE)

# Run GWAS.
gwasA <- gwas.asreml(
  pheno.data = gwas.data$pheno.data, resp = "Sucrose", gen = "Ind",
  fixedf = "Lots", residual = "Lots",
  Kinv = gwas.data$Kinv, Q = gwas.data$Q, npc = 3,
  geno.data = gwas.data$geno.data, map.data = gwas.data$map.data,
  pvalue.thr = 0.0005, bonferroni = FALSE, workspace = "2Gb")

# Simple Manhattan plot call.
manhattan.plot(gwas.table = gwasA$gwas.all)

# Manipulate some features.
manhattan.plot(
  gwas.table = gwasA$gwas.all,
  pvalue.thr = 0.0005, point.alpha = 0.80, padding = 3e6)

# Adding tags.
gwasA$gwas.sel$tag <-
  c("Putative gene 1", "Putative gene 2",
    "Putative gene 3", "Putative gene 4", "Putative gene 5")

manhattan.plot(
  gwas.table = gwasA$gwas.all,
  pvalue.thr = 0.0005, point.alpha = 0.80, padding = 3e6,
  tag.table = gwasA$gwas.sel, tag.repel = TRUE)
}
}
