% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map.plot}
\alias{map.plot}
\title{Draws the genetic map}
\usage{
map.plot(
  map.data = NULL,
  tag.markers = NULL,
  tag.repel = TRUE,
  tag.colour = "black",
  tag.size = 1.8,
  marker.colour = "#105E26",
  marker.width = 0.05,
  marker.alpha = 0.08,
  chrom.colour = "#DEDEDE",
  chrom.width = 4,
  chrom.contour = "#DEDEDE",
  message = TRUE
)
}
\arguments{
\item{map.data}{A data frame containing the map information (\code{"marker"}, \code{"chrom"},
and \code{"pos"}) (default = \code{NULL}).}

\item{tag.markers}{A (non-mandatory) vector with the identification (name) of markers to be highlighted
in the map (default = \code{NULL}).}

\item{tag.repel}{If \code{TRUE} tags are automatically repelled (default = \code{TRUE}).}

\item{tag.colour}{A character with a colour name (\emph{e.g.}, \code{"black"})
to be used in the colouring of selected markers labels (default = \code{"black"}).}

\item{tag.size}{A numeric value (>= 0) indicating label sizes for
selected marker names (default = \code{1.3}).}

\item{marker.colour}{A character with a colour name (\emph{e.g.}, \code{"blue"})
to be used in marker colouring (default = \code{"#105E26"}).}

\item{marker.width}{A numeric value (>= 0) indicating the marker width (represented by a line;
default = \code{0.05}).}

\item{marker.alpha}{A numeric value indicating the markers' transparency (default = \code{0.08}).}

\item{chrom.colour}{A character with a colour name (\emph{e.g.}, \code{"white"})
to be used in inner chromosome colouring (default = \code{"#DEDEDE"}).}

\item{chrom.width}{A numeric value (> 0) indicating the chromosome width (default = \code{4}).}

\item{chrom.contour}{A character with a colour name (\emph{e.g.} \code{"#DEDEDE"})
to be used in outer chromosome colouring (default = \code{"black"}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A ggplot object containing the genetic map.
}
\description{
Generates the genetic map for visualization.
}
\details{
This function was developed in collaboration with Khaled Al-Shamaa (ICARDA/CGIAR).
}
\examples{
\dontrun{
# Prepare Apricot dataset.
gwas.data <- pre.gwas(
  pheno.data = pheno.apricot, indiv = "Ind", resp = "Sucrose",
  geno.data = geno.apricot, map.data = map.apricot, maf = 0.05,
  marker.callrate = 0.40, impute = TRUE)

# Run GWAS.
gwasA <- gwas.asreml(
  pheno.data = gwas.data$pheno.data, resp = "Sucrose", gen = "Ind",
  fixedf = "Lots", residual = "Lots",
  Kinv = gwas.data$Kinv, Q = gwas.data$Q, npc = 3,
  geno.data = gwas.data$geno.data, map.data = gwas.data$map.data,
  pvalue.thr = 0.0005, bonferroni = FALSE, workspace = "2Gb")

# Get genetic map with the significant markers.
map.plot(map.data = gwasA$gwas.all,
  tag.markers = gwasA$gwas.sel$marker)

# Customizing a bit.
map.plot(map.data = gwasA$gwas.all,
  tag.markers = gwasA$gwas.sel$marker,
  marker.colour = "white", marker.width = 0.2,
  marker.alpha = 0.1, chrom.colour = "grey",
  chrom.contour = "grey", tag.colour = "#0072B2")
}
}
