% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_plot.R
\name{marker.plot}
\alias{marker.plot}
\title{Generates phenotype-by-genotype plots for selected markers}
\usage{
marker.plot(
  pheno.data = NULL,
  indiv = NULL,
  resp = NULL,
  geno.data.sel = NULL,
  type.plot = c("boxplot", "violin"),
  family = c("gaussian", "binomial"),
  sample.label = FALSE,
  sample.width = FALSE,
  maf.colour = FALSE,
  facet.dim = NULL,
  scales = NULL,
  message = TRUE
)
}
\arguments{
\item{pheno.data}{A data frame with all relevant columns (factors and covariates)
and phenotypic responses to be used (default = \code{NULL}).}

\item{indiv}{A character with the name of the column in \code{pheno.data}
with the identification of treatments (genotypes) (default = \code{NULL}).}

\item{resp}{A character with the name of the numeric variable in \code{pheno.data} with the
response variable (default = \code{NULL}).}

\item{geno.data.sel}{A matrix with SNP data of form \eqn{n \times s},
with \eqn{n} individuals and \eqn{s} selected markers. Individual and marker names are
assigned to \code{rownames} and \code{colnames}, respectively.
SNP data is coded as: 0, 1, 2 (default = \code{NULL}).}

\item{type.plot}{A character indicating the type of plot. Options are: \code{"boxplot"} and
\code{"violin"} plot (default = \code{"boxplot"}).}

\item{family}{A character indicating the family of the distribution for \code{resp}.
Options are: \code{"gaussian"} and \code{"binomial"} (default = \code{"gaussian"}).}

\item{sample.label}{If \code{TRUE} the sample size of each genotypic state is added to the \eqn{x}
axis (default = \code{FALSE}).}

\item{sample.width}{If \code{TRUE} the size of the boxplots will be proportional to sample size
(default = \code{FALSE}).}

\item{maf.colour}{If \code{TRUE} the plot colour will be proportional to the minor allele
frequency (default = \code{FALSE}).}

\item{facet.dim}{A numeric vector indicating the number of rows and columns to be used
in the faceting process, \emph{e.g.}, \code{c(4, 1)} for 4 rows and 1 column (default = \code{NULL}).}

\item{scales}{A character indicating if the scales/levels in \eqn{x} and \eqn{y} axis should be the
same values across facets (panels). Options are: \code{"fixed"} \code{"free_x"}, \code{"free_y"}
and \code{"free"}. See \link[ggplot2]{facet_wrap} for more information (default = \code{NULL}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A ggplot object containing one or more phenotype-by-genotype plots.
}
\description{
Generates a plot, based on a given set of markers, for visualizing the
phenotype in relation to the underlying genotypes.
Both Gaussian and binomial data are supported.
}
\details{
These plots are useful to verify and explore the validity of reported associations.
For example: 1) a linear change is expected on significant markers for additive action,
2) a non-linear pattern might indicate non-additive action,
3) genotype classes might differ on their phenotypic distribution (\emph{e.g.}, heterogeneity of variances), and
4) the influence of low MAF is clear and might indicate significant associations by chance.

Check provided references for examples of these plots and their interpretation.
}
\examples{
\dontrun{
# Prepare Apricot dataset.
gwas.data <- pre.gwas(
  pheno.data = pheno.apricot, indiv = "Ind", resp = "Sucrose",
  geno.data = geno.apricot, map.data = map.apricot, maf = 0.05,
  marker.callrate = 0.40, impute = TRUE)

# Run GWAS.
gwasA <- gwas.asreml(
  pheno.data = gwas.data$pheno.data, resp = "Sucrose", gen = "Ind",
  fixedf = "Lots", residual = "Lots",
  Kinv = gwas.data$Kinv, Q = gwas.data$Q, npc = 3,
  geno.data = gwas.data$geno.data, map.data = gwas.data$map.data,
  pvalue.thr = 0.0005, bonferroni = FALSE, workspace = "2Gb")

# Get name and data of significant markers.
sign.markers <- gwasA$gwas.sel$marker
geno.data.sel <- gwas.data$geno.data[, sign.markers]

# Simple marker plot call.
marker.plot(
 pheno.data = gwas.data$pheno.data, indiv = "Ind", resp = c("Sucrose"),
 geno.data.sel = geno.data.sel)

# Manipulate some features (including a trait).
marker.plot(
 pheno.data = gwas.data$pheno.data, indiv = "Ind",
 resp = c("Sucrose", "Ethylene"), geno.data.sel = geno.data.sel,
 sample.label = TRUE, sample.width = TRUE, maf.colour = TRUE,
 scales = "free")
}
}
\references{
Croteau-Chonka D.C., Rogers A.J., Raj T., McGeachie M.J., Qiu W., et al. 2015.
Expression quantitative trait loci information improves predictive modeling of
disease relevance of non-coding genetic variation.
\emph{PLOS ONE} 10(10):e0140758.

Galli G., Alves F.C., Morosini J.S., and Fritsche-Neto R. 2020.
On the usefulness of parental lines GWAS for predicting low heritability traits in tropical maize hybrids.
\emph{PLOS ONE} 15(2):e0228724.
}
