% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_position.R
\name{marker.position}
\alias{marker.position}
\title{Generates continuous relative position of markers}
\usage{
marker.position(
  map.data = NULL,
  chrom = "chrom",
  pos = "pos",
  collate = TRUE,
  padding = 0,
  message = TRUE
)
}
\arguments{
\item{map.data}{A data frame with each marker's name, chromosome, and position.
Variable names \strong{must} be "marker", "chrom", and "pos" (default = \code{NULL}).}

\item{collate}{If \code{TRUE} the chromosomes will be separated by one base pair (default = \code{TRUE}).}

\item{padding}{A \code{numeric} value that will be used as the distance between chromosomes (default = \code{0}).}
}
\value{
A vector with the continuous relative positions of markers.
}
\description{
Generates continuous relative position of markers
}
\examples{
map.data <- ASRgenomics:::dummy.map_(marker.id = 1:20)
map.data$chrom[11:20] <- 2
map.data$pos <- ASRgwas:::marker.position(
 map.data, collate = TRUE, padding = 5)
map.data
}
\keyword{internal}
