% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data.R
\name{match.data}
\alias{match.data}
\title{Check if objects to be used for GWAS model match}
\usage{
match.data(
  pheno.data = NULL,
  indiv = NULL,
  map.data = NULL,
  geno.data = NULL,
  Kinv = NULL,
  Q = NULL,
  caller = c("gwas", "gblup"),
  message = TRUE
)
}
\arguments{
\item{pheno.data}{A data frame with all relevant columns (factors and covariates)
and phenotypic responses to be used (default = \code{NULL}).}

\item{indiv}{A character with name of the column in \code{pheno.data}
with the genotypes (default = \code{NULL}).}

\item{map.data}{A data frame with each marker's name, chromosome, and position.
Variable names \strong{must} be "marker", "chrom", and "pos" (default = \code{NULL}).}

\item{geno.data}{A matrix with SNP data of form \eqn{n \times p},
with \eqn{n} individuals and \eqn{p} markers. Individual and marker names are
assigned to \code{rownames} and \code{colnames}, respectively.
SNP data is coded as: 0, 1, 2 (default = \code{NULL}).}

\item{Kinv}{A list of matrices representing the \strong{inverse} of the genomic relationship
matrix \eqn{\boldsymbol{Kinv}}. The matrices must be in sparse form (default = \code{NULL}).}

\item{Q}{A matrix of size \eqn{n} x \eqn{nv} with \eqn{nv} population structure-related covariates (\emph{e.g.},
principal component scores; default = \code{NULL}).}

\item{caller}{A character with the name of the function calling this function (default = \code{"gwas"}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A hard stop if any pair of the provided objects do not match.
}
\description{
Check if objects to be used for GWAS model match
}
\details{
The matching performed on non-null objects are:
(i) between \'rownames' in \'geno.data' if \'geno.data' is a list;
(ii) between \'colnames' in \'geno.data' if \'geno.data' is a list;
(iii) between the \'rowNames' attribute of \'Kinv' if \'Kinv' is a list;
(iv) between the \'colNames' attribute of \'Kinv' if \'Kinv' is a list;
(v) between the levels of \'pheno.data[['indiv']]' and the \'rownames' of \'geno.data',
\'rowNames' and \'colNames' of \'Kinv, and the \'rownames' of \'Q'.
(vi) between the \'colnames' of \'geno.data' and \'map.data[['marker']]'.
}
\keyword{internal}
