% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data.R
\name{match.in.list}
\alias{match.in.list}
\title{Match objects within a list (trap)}
\usage{
match.in.list(container = NULL, code = NULL)
}
\arguments{
\item{code}{The string of code to be used to check across objects (default = \code{NULL}).}

\item{list}{A list of objects to compare (default = \code{NULL}).}
}
\value{
A hard stop if a mismatch is found between any pair of objects.
}
\description{
Match objects within a list (trap)
}
\details{
The word "obj" is used in \code{code} to identify an
arbitrary object (each object in \code{list}).
}
\keyword{internal}
