% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_update.R
\name{mc.marker.update}
\alias{mc.marker.update}
\title{Implements GWAS model fit using \link[asreml]{update.asreml}}
\usage{
mc.marker.update(
  mod = NULL,
  geno.data = NULL,
  maxiter.update = NULL,
  threads = 1,
  obj.parallel = NULL,
  P3D = FALSE,
  message = TRUE
)
}
\arguments{
\item{mod}{A pre-fit \pkg{asreml} model object. Must follow the conditions of the function
(default = \code{NULL}).}

\item{geno.data}{A matrix with SNP data of form \eqn{n \times p},
with \eqn{n} individuals and \eqn{p} markers. Individual and marker names are
assigned to \code{rownames} and \code{colnames}, respectively.
SNP data is coded as: 0, 1, 2 (default = \code{NULL}).}

\item{maxiter.update}{A numeric value indicating the number of iterations to be
used in \link[asreml]{update.asreml} under the non-P3D method (default = \code{1}).}

\item{threads}{An integer with the number of threads to be used in parallel processing
(default = \code{1}).}

\item{obj.parallel}{A character vector indicating which objects passed to the function that must
be exported to clusters (\emph{e.g.}, objects passed to \code{Kinv} argument). If nothing is passed, the
algorithm will try to identify the required object. This is only required in Windows as Unix systems
share environments by default (default = \code{NULL}).}

\item{P3D}{If \code{TRUE} the "Population Parameters Previously Determined" algorithm will be used.
If \code{FALSE} the variance components are allowed to vary while fitting each marker individually.
(default = \code{FALSE}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A list with \code{effect}, \code{z.ratio} and \code{p.value} of each marker present
in \emph{geno.data}.
}
\description{
This functions runs conventional or P3D GWAS using \link[asreml]{update.asreml}.
One marker will be added to the original model at a time and the model updated.
}
\keyword{internal}
