% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{moore.penrose}
\alias{moore.penrose}
\title{Moore-Penrose pseudo-inverse}
\usage{
moore.penrose(G = NULL, eig.tol = 1e-06)
}
\arguments{
\item{G}{A matrix to be inverted (default = \code{NULL}).}

\item{eig.tol}{Defines relative relevance (\emph{i.e.}, non-zero) of eigenvalues
compared to the largest one. It determines which threshold of eigenvalues
will be treated as zero (default = \code{1e-06}).}
}
\value{
A Moore Penrose pseudo-inverse of G matrix.
}
\description{
Moore-Penrose pseudo-inverse
}
\keyword{internal}
