% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{obj.in.call_}
\alias{obj.in.call_}
\title{Collect object names from call.}
\usage{
obj.in.call_(call = NULL)
}
\arguments{
\item{call}{A model call (default = \code{NULL})}
}
\value{
A character vector with names of objects present in call.
}
\description{
Collect object names from call.
}
\keyword{internal}
