% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_P3D.R
\name{P3D}
\alias{P3D}
\title{Implements Population Parameters Previously Determined (P3D) algorithm}
\usage{
P3D(
  y = NULL,
  X = NULL,
  geno.data = NULL,
  V = NULL,
  nedf = NULL,
  inverse.update = 0,
  threads = 1,
  message = TRUE
)
}
\arguments{
\item{y}{A vector with complete (no missing) centred phenotypic values of length \eqn{n} (default = \code{NULL}).}

\item{X}{A design matrix of fixed effects of size \eqn{n \times f}, where \eqn{n} is the
number of samples (as in \code{y}) and \eqn{f} is the number of fixed effects to be fitted along
with markers (default = \code{NULL}).}

\item{geno.data}{A matrix with SNP data of form \eqn{n \times p},
with \eqn{n} individuals and \eqn{p} markers. Individual and marker names are
assigned to \code{rownames} and \code{colnames}, respectively.
SNP data is coded as: 0, 1, 2 (default = \code{NULL}).}

\item{V}{A matrix of size \eqn{n \times n} with the variance-covariance of \eqn{y} (default = \code{NULL}).}

\item{nedf}{The number of degrees of freedom from the base model. This is obtained from
\code{mod$nedf} (default = \code{NULL}).}

\item{inverse.update}{An integer indicating the number of Schulz-type iterative updates to be applied to
the inverse of \eqn{\boldsymbol{V}} = \emph{var(y)} if \emph{geno.data}
has missing values when running a Gaussian model under P3D.
The larger the value the better the approximation (under certain conditions).
If \code{inverse.update = -1} then the Woodbury's method is used instead.
If \code{inverse.update = 0} is used, then Schulz's procedure is performed with no iterations (no update),
been less precise but faster to run (default = \code{-1}).}

\item{threads}{An integer with the number of threads to be used in parallel processing
(default = \code{1}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A list with \code{effect}, \code{z.ratio} and \code{p.value} of each marker present
in \emph{geno.data}.
}
\description{
Implements Population Parameters Previously Determined (P3D) algorithm
}
\keyword{internal}
