% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_generic.R
\name{pre.generic}
\alias{pre.generic}
\title{Prepares and audits phenotypic and genomic data for GBLUP or GWAS}
\usage{
pre.generic(
  pheno.data = NULL,
  indiv = NULL,
  resp = NULL,
  weights = NULL,
  geno.data = NULL,
  map.data = NULL,
  rename.markers = TRUE,
  K = NULL,
  Q.method = c("none", "K", "geno.data"),
  return = c("pheno.data", "Kinv"),
  message = TRUE,
  caller = c("gwas", "gblup"),
  ellipsis = NULL
)
}
\arguments{
\item{pheno.data}{A mandatory data frame with all relevant columns (factors and
covariates) and one or more phenotypic responses to be used (default =
\code{NULL}).}

\item{indiv}{A mandatory character with the name of the column in \code{pheno.data}
with the identification of treatments (genotypes) (default = \code{NULL}).}

\item{resp}{A character with the name of the numeric variable in
\code{pheno.data} with the response variable (default = \code{NULL}).}

\item{weights}{A character with the name of the numeric variable in
\code{pheno.data} with the weights of each prediction (\emph{i.e.}, mean estimate)
(default = \code{NULL}).}

\item{geno.data}{A matrix with SNP data of form \eqn{n \times p}, with
\eqn{n} individuals and \eqn{p} markers. Individual and marker names are
assigned to \code{rownames} and \code{colnames}, respectively. SNP data is
coded as: 0, 1, 2 (default = \code{NULL}).}

\item{map.data}{An optional data frame with each marker's name, chromosome, and
position. Variable names \strong{must} be: "marker", "chrom", and "pos"
(default = \code{NULL}).}

\item{rename.markers}{If \code{TRUE} mathematical operators found on the
markers' names will be replaced with the character "_" (default =
\code{TRUE}).}

\item{K}{An optional symmetric matrix representing the genomic relationship
or kinship matrix \eqn{\boldsymbol{K}}. The matrix must be in full form
with size \eqn{n \times n}, with \eqn{n} individuals (genotypes). Genotypes
names must be present as row and column names. If not
provided, they will be obtained from \code{geno.data} (default =
\code{NULL}).}

\item{Q.method}{A character with the data source to be used for the
calculation of the \eqn{\boldsymbol{Q}} matrix by principal component
analyses (PCA). Options are \code{"K"} (for kinship) and \code{"geno.data"}
(for marker data) (default = \code{"K"}).}

\item{return}{A character vector specifying what to return. Options are:
\code{"pheno.data"}, \code{"geno.data"}, \code{"K"}, \code{"Kinv"},
\code{"map.data"}, \code{"Q"}, \code{"plot.scree"}, and
\code{"eigenvalues"} (default = \code{"Kinv"}).}

\item{message}{If \code{TRUE}, diagnostic messages are printed on screen
(default = \code{TRUE}).}

\item{ellipsis}{Further arguments to be called. Run \code{pre.gblup()} or
\code{pre.gwas()} to check possible arguments and their default values.
These arguments are called from functions:
\link[ASRgenomics]{qc.filtering}, \link[ASRgenomics]{snp.pruning},
\link[ASRgenomics]{G.matrix}, \link[ASRgenomics]{G.tuneup},
\link[ASRgenomics]{kinship.diagnostics}, \link[ASRgenomics]{G.inverse},
\link[ASRgenomics]{kinship.pca}, and \link[ASRgenomics]{snp.pca} available
from the package \pkg{ASRgenomics}.}
}
\value{
A list with several data frames and objects to be used in downstream
  GBLUP and GWAS model fits.

For GBLUP, these items are:

\itemize{
  \item \code{pheno.data}: A verified phenotypic data frame with all relevant columns (and rows) and responses to be used.
  \item \code{map.data}: A verified (or generated) data frame with markers' names, chromosomes, and positions.
  \item \code{geno.data}: A filtered and verified matrix with marker data.
  \item \code{K}: A list of verified and tuned-up kinship matrices.
  \item \code{Kinv}: A list of verified and tuned-up inverse kinship matrices in sparse format (ready for ASReml-R).
}

For GWAS, these items are:

\itemize{
  \item \code{pheno.data}: A verified phenotypic data frame with all relevant columns (and rows) and responses to be used.
  \item \code{map.data}: A verified (or generated) data frame with markers' names, chromosomes, and positions.
  \item \code{geno.data}: A filtered and verified matrix with marker data.
  \item \code{K}: A verified and tuned-up kinship matrix (if requested).
  \item \code{Kinv}: A verified and tuned-up inverse of the kinship matrix in sparse format (ready for ASReml-R).
  \item \code{Q}: A population structure \eqn{\boldsymbol{Q}} matrix of principal component scores.
  \item \code{plot.scree}: The scree plot for the components in the \eqn{\boldsymbol{Q}} matrix.
  \item \code{eigenvalues}: The eigenvalues for the components in the \eqn{\boldsymbol{Q}} matrix.
}
}
\description{
This function checks the phenotypic and genomic data and to
  prepare the required data frames for downstream GBLUP or GWAS with
  \code{gblup.asreml} and \code{gwas.asreml}, respectively.
  The mandatory input is \code{pheno.data},
  \code{indiv}, and \code{geno.data}. If not provided, the map, and the
  matrices \eqn{\boldsymbol{K}}, \eqn{\boldsymbol{Kinv}} and
  \eqn{\boldsymbol{Q}} will be generated.
}
\details{
A summary of the workflow considered in \code{pre.gblup} is:
\itemize{
  \item 1: Rename markers (if required).
  \item 2: Organize variables/columns in \code{pheno.data} (if necessary).
  \item 3: Sort provided data alphanumerically based on genotypes/individuals names.
  \item 4: Pre-match \code{pheno.data} and \code{geno.data} if the latter was provided.
  \item 5: Run \link[ASRgenomics]{qc.filtering} on \code{geno.data} to obtain filtered marker matrix.
  \item 6: Round genotypic values on \code{geno.data} (if decimals are present) as required by \link[ASRgenomics]{G.matrix}.
  \item 7: Obtain matrix \eqn{\boldsymbol{K}} using function \link[ASRgenomics]{G.matrix}, or use the one provided by user.
  \item 8: Run \link[ASRgenomics]{G.tuneup} on \eqn{\boldsymbol{K}} with blending, bending, and aligning (if requested by user).
  \item 9: Perform diagnostics on \eqn{\boldsymbol{K}} matrix using \link[ASRgenomics]{kinship.diagnostics}.
  \item 10: Repeat match of all datasets after \eqn{\boldsymbol{K}} quality control.
  \item 11: Reorder all matrices based on \code{levels(pheno.data[[indiv]])}.
  \item 12: Obtain inverse of kinship matrix (\eqn{\boldsymbol{Kinv}}) using \link[ASRgenomics]{G.inverse}.
  \item 13: If \eqn{\boldsymbol{Kinv}} is ill-conditioned (changes are cumulative) then:
    \itemize{
      \item 13.1: Run \link[ASRgenomics]{G.tuneup} on \eqn{\boldsymbol{K}} with blending of 0.05 using an identity matrix. If inverse is still ill-conditioned then go to 13.2; otherwise go to 14.
      \item 13.2: Run \link[ASRgenomics]{G.tuneup} on \eqn{\boldsymbol{K}} again with blending of 0.10 using an identity matrix. If inverse is still ill-conditioned then stop; otherwise go to 14.
    }
  \item 14: Return requested objects.
}

A summary of the workflow considered in \code{pre.gwas} is:
\itemize{
  \item 1: Rename markers (if required).
  \item 2: Remove individual records with \code{NA} on \code{resp} and \code{weights}.
  \item 3: Organize variables/columns in \code{pheno.data} (if necessary).
  \item 4: Pre-match \code{pheno.data} and \code{geno.data}.
  \item 5: Run \link[ASRgenomics]{qc.filtering} on \code{geno.data} to obtain filtered marker matrix.
  \item 6: Round genotypic values on \code{geno.data} (if decimals are present) as required by \link[ASRgenomics]{G.matrix}.
  \item 7: Obtain matrix \eqn{\boldsymbol{K}} using function \link[ASRgenomics]{G.matrix}, or use the one provided by user.
  \item 8: Run \link[ASRgenomics]{G.tuneup} on \eqn{\boldsymbol{K}} with blending, bending, and aligning (if requested by user).
  \item 9: Perform diagnostics on \eqn{\boldsymbol{K}} matrix using \link[ASRgenomics]{kinship.diagnostics}.
  \item 10: Repeat match of all datasets after \eqn{\boldsymbol{K}} quality control.
  \item 11: Reorder all matrices based on \code{levels(pheno.data[[indiv]])}.
  \item 12: Obtain inverse of kinship matrix (\eqn{\boldsymbol{Kinv}}) using \link[ASRgenomics]{G.inverse}.
  \item 13: If \eqn{\boldsymbol{Kinv}} is ill-conditioned (changes are cumulative) then:
    \itemize{
      \item 13.1: Run \link[ASRgenomics]{G.tuneup} on \eqn{\boldsymbol{K}} with blending of 0.05 using an identity matrix. If inverse is still ill-conditioned then go to 13.2; otherwise go to 14.
      \item 13.2: Run \link[ASRgenomics]{G.tuneup} on \eqn{\boldsymbol{K}} again with blending of 0.10 using an identity matrix. If inverse is still ill-conditioned then stop; otherwise go to 14.
    }
  \item 14: Obtain matrix \eqn{\boldsymbol{Q}} based on \code{geno.data} or final \eqn{\boldsymbol{K}} matrix.
}

The majority of the default values used for arguments are those suggested
in \pkg{ASRgenomics}, but these can be modified via the ellipsis argument (\code{...}).
Nevertheless, some defaults have been changed for better implementation of
GBLUP and GWAS downstream procedures.

The modified defaults for GBLUP are: \code{ncp = 20}, \code{maf = 0.05},
\code{marker.callrate = 0.2}, \code{ind.callrate = 0.2}, \code{clean.diagonal
= TRUE}, \code{clean.duplicate = TRUE}, \code{diagonal.thr.large = 1.8}, and
\code{diagonal.thr.small = 0.2}. It might be important to adjust these
parameters along with GBLUP runs to identify the best analytical approach.

The modified defaults for GWAS are: \code{ncp = 20}, \code{maf = 0.05},
\code{marker.callrate = 0.2}, \code{ind.callrate = 0.2}, \code{clean.diagonal
= TRUE}, \code{clean.duplicate = TRUE}, \code{diagonal.thr.large = 1.8}, and
\code{diagonal.thr.small = 0.2}. It might be important to adjust these
parameters along with GWAS runs to identify best analytical approach.

For GWAS, the \eqn{\boldsymbol{Q}} matrix can be calculated based on
\eqn{\boldsymbol{K}} or \code{geno.data}. Nevertheless, if missing values are
still present in \code{geno.data} after cleaning, \code{Q.method = "K"} will
be always used, as this method allows for missing values.

For more information on the aforementioned functions, please refer to the
\pkg{ASRgenomics} help.
}
\keyword{internal}
