% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq.plot}
\alias{qq.plot}
\title{Draws one or more Q-Q plots for GWAS marker data results}
\usage{
qq.plot(
  gwas.table = NULL,
  point.colour = "#0072B2",
  point.size = 1,
  point.alpha = 1,
  trend.colour = "grey",
  coord.equal = TRUE,
  gwas.index = NULL,
  facet.main = NULL,
  facet.secondary = NULL,
  facet.dim = NULL,
  legend.position = "none",
  message = TRUE
)
}
\arguments{
\item{gwas.table}{A data frame of one or more (vertically stacked) GWAS analyses with (as a minimum)
the \code{"p.value"} column of marker results (default = \code{NULL}).}

\item{point.colour}{A character with a colour name (\emph{e.g.}, \code{"blue"}) or a column name
from \code{gwas.table} to be used as the index for point colouring (default = \code{"#0072B2"}).}

\item{point.size}{A numeric value (greater than 0) indicating the point size
(default = \code{1}).}

\item{point.alpha}{A numeric value indicating the points' transparency (default = \code{1}).}

\item{trend.colour}{A character with a colour name (\emph{e.g.}, \code{"blue"}) to be used
for colouring of the central line (expected value) (default = \code{"black"}).}

\item{coord.equal}{If \code{TRUE} the \eqn{x} and \eqn{y} coordinates will have the same limits
and a ratio of 1 (default = \code{TRUE}).}

\item{gwas.index}{A character indicating a column name in \code{gwas.table} identifying each GWAS analysis.
This is only necessary when more than one GWAS analysis is present in \code{gwas.table} (default = \code{NULL}).}

\item{facet.main}{A character indicating a column name in \code{gwas.table} to be used
for the main faceting of panels (if more than one GWAS analysis is present) (default = \code{NULL}).}

\item{facet.secondary}{A character indicating a column name in \code{gwas.table} to be used
as secondary faceting of panels (if more than one GWAS analysis is present) (default = \code{NULL}).}

\item{facet.dim}{A numeric vector indicating the number of rows and columns to be used
in the faceting process, \emph{e.g.}, \code{c(4, 1)}, for 4 rows and 1 column. This argument only works
if \code{facet.main} is provided, and works best if \code{facet.secondary} is \code{NULL}
(default = \code{NULL}).}

\item{legend.position}{A character indicating the position for the legend in the displayed plot.
Options are: \code{"bottom"}, \code{"top"}, \code{"left"}, \code{"right"}, \code{"none"}, or the
corresponding vector of coordinates (such as \code{c(0.95, 0.05)} for \eqn{x} and \eqn{y}, respectively)
(default = \code{"none"}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A ggplot objects containing one or more Q-Q plots.
}
\description{
Generates one or more quantile-quantile (Q-Q) plots for visualization.
}
\examples{
\dontrun{
# Prepare Apricot dataset.
gwas.data <- pre.gwas(
  pheno.data = pheno.apricot, indiv = "Ind", resp = "Sucrose",
  geno.data = geno.apricot, map.data = map.apricot, maf = 0.05,
  marker.callrate = 0.40, impute = TRUE)

# Run GWAS.
gwasA <- gwas.asreml(
  pheno.data = gwas.data$pheno.data, resp = "Sucrose", gen = "Ind",
  fixedf = "Lots", residual = "Lots",
  Kinv = gwas.data$Kinv, Q = gwas.data$Q, npc = 3,
  geno.data = gwas.data$geno.data, map.data = gwas.data$map.data,
  pvalue.thr = 0.0005, workspace = "2Gb")

# Simple QQ plot call.
qq.plot(gwas.table = gwasA$gwas.all)

# Manipulate some features.
qq.plot(
  gwas.table = gwasA$gwas.all, point.colour = "green",
  point.alpha = 0.5, point.size = 2)
}
}
