% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_aux.R
\name{r.dsum}
\alias{r.dsum}
\title{Direct sum of matrices [adapted from asremlPlus]}
\usage{
r.dsum(matrices)
}
\arguments{
\item{matrices}{List of matrices to be combined via direct sum (no default).}
}
\value{
The direct sum of the provided matrices.
}
\description{
Direct sum of matrices [adapted from asremlPlus]
}
\author{
Chris Brien; VSNi
}
\keyword{internal}
