% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatability_pev.R
\name{repeatability.pev}
\alias{repeatability.pev}
\title{Calculates repeatability based on the PEV}
\usage{
repeatability.pev(mod = NULL, component = NULL)
}
\arguments{
\item{mod}{An object of class \code{asreml} (default = \code{NULL}).}

\item{component}{The name of the variance component to be used in the PEV calculations
(\emph{e.g.}, "vm(fishid, Kinv$fishid)") (default = \code{NULL}).}
}
\value{
The repeatability of the requested variance component based on PEV.
}
\description{
Internal routine that calculates the repeatability based on
the Prediction Error Variance (PEV).
}
\details{
Calculations are done using:

\deqn{repeatability = 1 - mean(std.error^2) / \sigma^2_c}

where: \eqn{std.error} is the average of the standard errors (SE) of the requested factor's
solution and \eqn{\sigma^2_c} is its variance component estimate.
}
\keyword{internal}
