% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatability_vc.R
\name{repeatability.vc}
\alias{repeatability.vc}
\title{Calculates repeatability based on variance components}
\usage{
repeatability.vc(
  mod = NULL,
  numerator = NULL,
  bound.exclusions = NULL,
  name = "repeatability"
)
}
\arguments{
\item{mod}{An object of class \code{asreml} (default = \code{NULL}).}

\item{numerator}{The name of the variance component(s) to be used in the numerator of the
calculations (\emph{e.g.}, "vm(fishid, Kinv$fishid)") (default = \code{NULL}).}

\item{bound.exclusions}{The code of the boundary to be excluded (\emph{e.g.}, c("F", "B");
default = \code{NULL})}
}
\value{
A data frame with the estimates, standard error, and formula for the h2.vc calculation.
}
\description{
Uses \link[asreml]{vpredict} to calculate the repeatability
based on the variance components for a given (set of) component(s).
}
\details{
The algorithm uses the values provided in \code{numerator} as the numerator in the formula.
All variance components are added to the denominator, except when its bound conditions match values
passed to \code{bound.exclusions}.

The formula can be represented as:

\deqn{repeatability = \sigma^2_n / \sigma^2_d}

where: \eqn{\sigma^2_n} is the sum of the variance components passed to the \code{numerator} argument,
and \eqn{\sigma^2_d} is the sum of all variance components.

If there are heterogeneous residual variances, these will be averaged before calculations.
}
\keyword{internal}
