% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatability_wrap.R
\name{repeatability.wrap}
\alias{repeatability.wrap}
\title{An internal wrapper to calculate VC and PEV heritability/repeatability}
\usage{
repeatability.wrap(mod = NULL, herit.numerators = NULL, skip.vc = FALSE)
}
\arguments{
\item{mod}{An object of class \code{asreml} (default = \code{NULL}).}

\item{herit.numerators}{The name of the variance component(s) to be used in the numerator of the
calculations (\emph{e.g.}, "vm(fishid, Kinv$fishid)") (default = \code{NULL}).}

\item{skip.vc}{If FALSE, heritability based on variance components
will not be calculated (default = \code{FALSE}).}
}
\value{
The repeatabilities based on the variance components and prediction
error variances.
}
\description{
An internal wrapper to calculate VC and PEV heritability/repeatability
}
\keyword{internal}
