% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{schulz}
\alias{schulz}
\title{Performs Schulz-type inverted matrix update for row/column elimination}
\usage{
schulz(X = NULL, Xinv.init = NULL, na = NULL, niter = 2)
}
\arguments{
\item{X}{A matrix to be inverted (default = \code{NULL}).}

\item{Xinv.init}{An initial guess of the updated inverse (default = \code{NULL}).}

\item{na}{A vector indicating the indices of rows/columns to be dropped on \code{X}
(default = \code{NULL}).}

\item{niter}{An integer indicating the number of iterations to carry out
(theoretically, the more iterations, the closer the approximated inverse is to the true inverse)
(default = \code{2})}
}
\value{
The updated matrix inverse with 0 on dropped rows/columns.
}
\description{
Performs Schulz-type inverted matrix update for row/column elimination
}
\keyword{internal}
