% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_marker.R
\name{select.marker}
\alias{select.marker}
\title{Selects final set of GWAS markers using backward elimination}
\usage{
select.marker(
  gwas.object = NULL,
  geno.data.sel = NULL,
  ref.vc = 1,
  pvalue.thr = 0.1,
  maxiter.update = 5,
  workspace = "1Gb",
  message = TRUE
)
}
\arguments{
\item{gwas.object}{A (list) with all objects obtained
after running the function \link{gwas.asreml} (default = \code{NULL}).}

\item{geno.data.sel}{A matrix with (a subset of) marker data of form \eqn{n \times s}, with \eqn{n} individuals and
\eqn{s} "significant" markers to be evaluated in the backward selection process. Individuals and marker names
are assigned to \code{rownames} and \code{colnames}, respectively.
Markers can have any coding (\emph{e.g.}, additive or dominant) as they will be fitted as covariates in the model (default = \code{NULL}).}

\item{ref.vc}{A numeric (vector) containing the indices (or positions) of the variance components
(from \code{gwas.object$mod$vparameters}) to be used as the reference for the variance explained
by the markers (usually the genetic variance) (default = \code{1}).}

\item{pvalue.thr}{A numeric value with the corresponding \eqn{p-value}
threshold to identify significant markers for the elimination process.
Markers will be eliminated until all have a \eqn{p-value} smaller
than the value provided (default = \code{0.1})}

\item{maxiter.update}{A numeric value indicating the number of iterations to be
used in \link[asreml]{update.asreml} (default = \code{3L}).}

\item{workspace}{Specifies the workspace to be used by the \code{asreml} function
(default = \code{"1Gb"}).}

\item{message}{If \code{TRUE} diagnostic messages are printed on screen (default = \code{TRUE}).}
}
\value{
A list containing:
\itemize{
  \item \code{call}: An object of class \code{call} with the code used to fit \code{mod}.
  \item \code{gwas.sel}: A vector with the names of the final marker set with \code{p-value < pvalue.thr}.
  \item \code{mod}: The GWAS model fitted with the retained markers. Contains the same set of columns as \code{gwas.sel} from \link{gwas.asreml}.
  \item \code{aov}: The ANOVA table of the updated model with the final marker set.
  \item \code{expl.var}: The percentage of the genetic variance explained by the final marker set.
  \item \code{collinear.markers} [returned if collinearity is an issue]: A data frame with the names and correlation values of highly collinear/correlated markers that might affect model fitting.
}
}
\description{
The backward elimination technique starts by adding all markers to the base model as covariates
and obtaining their \eqn{p-value}. The marker with the highest \eqn{p-value} is eliminated and
the model is re-fit. The process is repeated until all retained markers have a \eqn{p-value}
smaller than the provided threshold.
}
\details{
If there is a group of markers with high collinearity (\emph{i.e.}, correlation >= 0.95)
then the process stops and the list of markers that are highly correlated
is reported. Here, one or more of these markers needs to be eliminated
for the backward elimination to proceed.
}
\examples{
\dontrun{
# Prepare Apricot dataset.
gwas.data <- pre.gwas(
  pheno.data = pheno.apricot, indiv = "Ind", resp = "Sucrose",
  geno.data = geno.apricot, map.data = map.apricot, maf = 0.05,
  marker.callrate = 0.40, impute = TRUE)

# Run GWAS.
gwasA <- gwas.asreml(
  pheno.data = gwas.data$pheno.data, resp = "Sucrose", gen = "Ind",
  fixedf = "Lots", residual = "Lots",
  Kinv = gwas.data$Kinv, Q = gwas.data$Q, npc = 3,
  geno.data = gwas.data$geno.data, map.data = gwas.data$map.data,
  pvalue.thr = 0.0005, bonferroni = FALSE, workspace = "2Gb")

# Get name of significant markers.
sign.markers <- gwasA$gwas.sel$marker
geno.data.sel <- gwas.data$geno.data[, sign.markers]

# Identify the index of genetic variance components.
gwasA$mod$vparameters

# Run of marker selection (this will return only the correlations).
set <- select.marker(
  gwas.object = gwasA, geno.data.sel = geno.data.sel,
  ref.vc = 1, pvalue.thr = 0.1)
set$call
set$aov
}
}
