% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_asreml.R
\name{v.asreml}
\alias{v.asreml}
\title{Calculates the variance-covariance of \eqn{y} (\eqn{\strong{V}} matrix) [adapted from asremlPlus]}
\usage{
v.asreml(
  asreml.obj = NULL,
  which.matrix = c("V", "G", "R"),
  Ks = NULL,
  weights = NULL
)
}
\arguments{
\item{asreml.obj}{An \pkg{asreml} object (default = \code{NULL}).}

\item{which.matrix}{This will probably be removed.}

\item{Ks}{A list with (non-inverted) full relationship matrices. See details (default = \code{NULL}).}

\item{weights}{A character indicating a variable in \code{asreml.obj} containing
residual weights (default = \code{NULL}).}
}
\value{
The variance-covariance of \eqn{y}: \eqn{\strong{V}} matrix.
}
\description{
This function calculates variance-covariance of \eqn{y} based on an object of class \code{asreml}
(fitted model).

We are not using asremlPlus here because we are allowing interactions with vms(). It would be
difficult to get the additional matrices outside the function. Lots of things have been commented
out but kept here for future reference and comparison to the main function.
}
\details{
Changes from the original function (asremlPlus):

* \code{vm} structure has been added;
* interactions/nesting with \code{vm} has been added;
* identification of \code{gammaPar} has changed;
* added possibility to include residual weights in the calculations;
* several things have been switched off for now;

Terminology:

\code{term}: a composite model term; e.g. \code{vm(ID, Kinv):Env}

A \code{ranterm} is anything passed to \code{random=~} in \pkg{asreml} (e.g. \code{genotype:env}).
This is later split into vars (e.g. if term is \code{genoype:env}, we will get
\code{genotype} and \code{env} as vars).

\code{termvars}: are the vars in each term. Might be one or more.
Argument \code{Ks} should be passed as follows: assuming \eqn{\strong{KA}} and \eqn{\strong{KD}} are two
relationship matrices inverted to \eqn{\strong{KinvA}} and \eqn{\strong{KinvD}}, respectively, which were used as
\code{vm} structures for fitting \code{asreml.obj}; the \code{Ks} should be
\code{list("vm(genA, KinvA)" = KA, "vm(genD, KinvD)" = KD)}, where \code{vm(genA, KinvA)} and
\code{vm(genD, KinvD)} are the names of the random terms; and \code{KA} and \code{KD}
are the original relationship matrices.

Additional information:

* Structures accepted for the random terms are: \code{id}, \code{idv}, and \code{vm}.
More can be added later on.
* Structures accepted for the residual terms are: \code{id}, \code{idv}, and \code{dsum}.
More can be added later on.
* Interaction/nesting with two \code{vm} is not permitted by \pkg{asreml} (too complex).
* It iterates through each term's variable (factor).
At each run, the Kronecker will be calculated with the previous terms.
This is used when there are interactions (more than one variable in the term).
* G.param has information about each term. Inside each term we have a \code{$varinace}
object and \code{$-name of var-} objects. The variance of the term might be on
\code{$varinace$initial} or in one of the vars \code{$-name of var-$initial}.
We also have to collect the \code{$model} associated to the object. If \code{model == "idv"};
the \code{$initial} will have the variance component. If not \code{"idv"}
it will be a correlation matrix, e.g. \code{vm}, \code{id}, etc.
* When \code{vm} is used, the relationship matrix is the correlation matrix itself.
It can be directly multiplied by the associated variance to get \eqn{\strong{G}}.
* There is a loop to sum up the variance of all random terms, and get \eqn{\strong{G}}.
\eqn{\strong{G}} (variance-covariance of random terms) and \eqn{\strong{R}}
(variance-covariance of residuals) are estimated separately and then summed up.
* If \code{gammaPar} has been used, the \eqn{\strong{V}} matrix is multiplied by the
scaling factor \code{$sigma2} at the end.
}
\author{
Chris Brien; VSNi
}
\keyword{internal}
