% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{woodbury}
\alias{woodbury}
\title{Performs Woodbury's inverted matrix update for row/column elimination}
\usage{
woodbury(X, na)
}
\arguments{
\item{X}{An inverted matrix to be updated (default = \code{NULL}).}

\item{na}{A vector indicating the indices of rows/columns to be dropped (default = \code{NULL}).}
}
\value{
The updated matrix inverse without dropped rows/columns.
}
\description{
Performs Woodbury's inverted matrix update for row/column elimination
}
\keyword{internal}
