% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASRgwas-package.R
\docType{package}
\name{ASRgwas-package}
\alias{ASRgwas}
\alias{ASRgwas-package}
\title{ASRgwas: ASRgwas an R package to perform Genome-Wide Association Studies (GWAS)}
\description{
ASRgwas is a comprehensive R package designed to perform Genome-Wide Association Studies (GWAS) using the modelling flexibility available in ASReml-R. This library assists with preparing and auditing phenotypic and genomic data; fitting GWAS models and identifying significant markers; and evaluating and exploring output.
}
\author{
\strong{Maintainer}: Salvador Gezan \email{salvador.gezan@vsni.co.uk}

Authors:
\itemize{
  \item Salvador Gezan \email{salvador.gezan@vsni.co.uk}
  \item Giovanni Galli \email{giovanni.galli@vsni.co.uk}
  \item Didier Murillo \email{didier.murilloflorez@ndus.edu}
  \item Darren Murray \email{darren.murray@vsni.co.uk}
}

Other contributors:
\itemize{
  \item VSN International [copyright holder]
}

}
\keyword{internal}
